Update Faiss to 1.7.0

+++ src/main/java/com/vectorsearch/faiss/swig/IndexBinaryFlat.java
@@ -60,11 +60,11 @@ public class IndexBinaryFlat extends IndexBinary {
     return swigfaissJNI.IndexBinaryFlat_query_batch_size_get(swigCPtr, this);
   }
 
-  public IndexBinaryFlat(int d) {
+  public IndexBinaryFlat(long d) {
     this(swigfaissJNI.new_IndexBinaryFlat__SWIG_0(d), true);
   }
 
-  public void add(int n, SWIGTYPE_p_unsigned_char x) {
+  public void add(long n, SWIGTYPE_p_unsigned_char x) {
     swigfaissJNI.IndexBinaryFlat_add(swigCPtr, this, n, SWIGTYPE_p_unsigned_char.getCPtr(x));
   }
 
@@ -72,15 +72,15 @@ public class IndexBinaryFlat extends IndexBinary {
     swigfaissJNI.IndexBinaryFlat_reset(swigCPtr, this);
   }
 
-  public void search(int n, SWIGTYPE_p_unsigned_char x, int k, SWIGTYPE_p_int distances, SWIGTYPE_p_long labels) {
+  public void search(long n, SWIGTYPE_p_unsigned_char x, long k, SWIGTYPE_p_int distances, SWIGTYPE_p_long labels) {
     swigfaissJNI.IndexBinaryFlat_search(swigCPtr, this, n, SWIGTYPE_p_unsigned_char.getCPtr(x), k, SWIGTYPE_p_int.getCPtr(distances), SWIGTYPE_p_long.getCPtr(labels));
   }
 
-  public void range_search(int n, SWIGTYPE_p_unsigned_char x, int radius, RangeSearchResult result) {
+  public void range_search(long n, SWIGTYPE_p_unsigned_char x, int radius, RangeSearchResult result) {
     swigfaissJNI.IndexBinaryFlat_range_search(swigCPtr, this, n, SWIGTYPE_p_unsigned_char.getCPtr(x), radius, RangeSearchResult.getCPtr(result), result);
   }
 
-  public void reconstruct(int key, SWIGTYPE_p_unsigned_char recons) {
+  public void reconstruct(long key, SWIGTYPE_p_unsigned_char recons) {
     swigfaissJNI.IndexBinaryFlat_reconstruct(swigCPtr, this, key, SWIGTYPE_p_unsigned_char.getCPtr(recons));
   }
