+++ src/main/java/com/vectorsearch/faiss/swig/NormalizationTransform.java
@@ -55,11 +55,11 @@ public class NormalizationTransform extends VectorTransform {
     this(swigfaissJNI.new_NormalizationTransform__SWIG_2(), true);
   }
 
-  public void apply_noalloc(int n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
+  public void apply_noalloc(long n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
     swigfaissJNI.NormalizationTransform_apply_noalloc(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(x), SWIGTYPE_p_float.getCPtr(xt));
   }
 
-  public void reverse_transform(int n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
+  public void reverse_transform(long n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
     swigfaissJNI.NormalizationTransform_reverse_transform(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(xt), SWIGTYPE_p_float.getCPtr(x));
   }
