+++ src/main/java/com/vectorsearch/faiss/swig/PCAMatrix.java
@@ -114,7 +114,7 @@ public class PCAMatrix extends LinearTransform {
     this(swigfaissJNI.new_PCAMatrix__SWIG_4(), true);
   }
 
-  public void train(int n, SWIGTYPE_p_float x) {
+  public void train(long n, SWIGTYPE_p_float x) {
     swigfaissJNI.PCAMatrix_train(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(x));
   }
