+++ src/main/java/com/vectorsearch/faiss/swig/RemapDimensionsTransform.java
@@ -56,11 +56,11 @@ public class RemapDimensionsTransform extends VectorTransform {
     this(swigfaissJNI.new_RemapDimensionsTransform__SWIG_2(d_in, d_out), true);
   }
 
-  public void apply_noalloc(int n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
+  public void apply_noalloc(long n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
     swigfaissJNI.RemapDimensionsTransform_apply_noalloc(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(x), SWIGTYPE_p_float.getCPtr(xt));
   }
 
-  public void reverse_transform(int n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
+  public void reverse_transform(long n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
     swigfaissJNI.RemapDimensionsTransform_reverse_transform(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(xt), SWIGTYPE_p_float.getCPtr(x));
   }
