+++ src/test/java/com/vectorsearch/faiss/ThreeIVFPQ.java
@@ -37,7 +37,7 @@ public class ThreeIVFPQ extends FaissTestCase {
         index.train(inputRowCount, inputVectors.cast());
         index.add(inputRowCount, inputVectors.cast());
         final boolean isTrained = index.getIs_trained();
-        final int nTotal = index.getNtotal();
+        final long nTotal = index.getNtotal();
         final String msg = "isTrained = " + isTrained + ", nTotal = " + nTotal;
         LOGGER.info(msg);
     }
